Added Picker for selecting city
Instead of a Entry to fill in the city you want to go. I've added a picker wich shows all the cities using the apiCaller. This is more userfriendly since the user will no know witch cites there are. For some reason it won't display on the right place in the gui, so its in the top corner